+++ src/pages/order-confirm/order-content/CouponModalButton.tsx
@@ -1,7 +1,7 @@
 import { BasicModal } from "@dev-dino22/modal-components";
-import Button from "../../components/common/Button";
+import Button from "../../../components/common/Button";
 import { useState } from "react";
-import CouponModalContent from "./coupon-modal/CouponModalContent";
+import CouponModalContent from "../coupon-modal/CouponModalContent";
 
 function CouponModalButton() {
   const [isOpen, setIsOpen] = useState(false);
