+++ src/pages/order-confirm/order-content/OrderList.tsx
@@ -1,9 +1,9 @@
 import styled from "@emotion/styled";
-import { getShoppingCartData } from "../../api/cart";
-import { useAPIDataContext } from "../../context/APIDataProvider";
-import { useOrderListContext } from "../../context/OrderListProvider";
-import Image from "../../components/common/Image";
-import { formatKRWString } from "../../utils/formatKRWString";
+import { getShoppingCartData } from "../../../api/cart";
+import { useAPIDataContext } from "../../../context/APIDataProvider";
+import { useOrderListContext } from "../../../context/OrderListProvider";
+import Image from "../../../components/common/Image";
+import { formatKRWString } from "../../../utils/formatKRWString";
 
 function OrderList() {
   const { data: cartListData } = useAPIDataContext({
